# Conflicts:
#	Admin/src/main/java/com/example/admin/Controller/Staff/StaffHomeController.java
#	Admin/src/main/java/com/example/admin/Converter/ProductConverter.java
#	Admin/src/main/java/com/example/admin/Domain/Product.java
#	Admin/src/main/java/com/example/admin/Repository/ProductRepository.java
#	Admin/src/main/java/com/example/admin/Service/Impl/ProductServiceImpl.java
#	Admin/src/main/java/com/example/admin/Service/ProductService.java
#	Admin/src/main/resources/application.properties
#	Admin/target/classes/application.properties
#	Admin/target/classes/com/example/admin/Controller/Staff/StaffHomeController.class
#	Admin/target/classes/com/example/admin/Converter/ProductConverter.class
#	Admin/target/classes/com/example/admin/Domain/Product.class

+++ Admin/src/main/java/com/example/admin/Entity/ProductEntity.java
@@ -32,7 +32,7 @@ public class ProductEntity {
 
     private String sub_info;
 
-    private int overall_rating;
+    private double overall_rating;
 
     private double discount;
